Lower MSRV and update to automatically feature gate instrinics.
