+++ devel/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "float16-benchmark"
+version = "0.0.1"
+authors = ["Alex Huszagh <ahuszagh@gmail.com>"]
+edition = "2021"
+publish = false
+
+[dependencies.float16]
+path = ".."
+default-features = false
+features = []
+
+[dev-dependencies]
+criterion = { version = "0.5", features = ["html_reports"] }
+quickcheck = "1.0"
+quickcheck_macros = "1.0"
+
+[[bench]]
+name = "convert"
+path = "convert.rs"
+harness = false
