+++ devel/src/lib.rs
@@ -1,0 +1,1 @@
+// Empty just for it to be a library.
