+++ src/binary16/arch/aarch64.rs
@@ -1,3 +1,4 @@
+#[rustversion::since(1.59)]
 use core::{
     arch::{
         aarch64::{float32x4_t, float64x2_t, uint16x4_t},
@@ -7,8 +8,9 @@ use core::{
     ptr,
 };
 
-#[target_feature(enable = "fp16")]
 #[inline]
+#[rustversion::since(1.59)]
+#[target_feature(enable = "fp16")]
 pub(super) unsafe fn f16_to_f32_fp16(i: u16) -> f32 {
     let result: f32;
     asm!(
@@ -19,8 +21,9 @@ pub(super) unsafe fn f16_to_f32_fp16(i: u16) -> f32 {
     result
 }
 
-#[target_feature(enable = "fp16")]
 #[inline]
+#[rustversion::since(1.59)]
+#[target_feature(enable = "fp16")]
 pub(super) unsafe fn f16_to_f64_fp16(i: u16) -> f64 {
     let result: f64;
     asm!(
@@ -31,8 +34,9 @@ pub(super) unsafe fn f16_to_f64_fp16(i: u16) -> f64 {
     result
 }
 
-#[target_feature(enable = "fp16")]
 #[inline]
+#[rustversion::since(1.59)]
+#[target_feature(enable = "fp16")]
 pub(super) unsafe fn f32_to_f16_fp16(f: f32) -> u16 {
     let result: u16;
     asm!(
@@ -43,8 +47,9 @@ pub(super) unsafe fn f32_to_f16_fp16(f: f32) -> u16 {
     result
 }
 
-#[target_feature(enable = "fp16")]
 #[inline]
+#[rustversion::since(1.59)]
+#[target_feature(enable = "fp16")]
 pub(super) unsafe fn f64_to_f16_fp16(f: f64) -> u16 {
     let result: u16;
     asm!(
@@ -55,8 +60,9 @@ pub(super) unsafe fn f64_to_f16_fp16(f: f64) -> u16 {
     result
 }
 
-#[target_feature(enable = "fp16")]
 #[inline]
+#[rustversion::since(1.59)]
+#[target_feature(enable = "fp16")]
 pub(super) unsafe fn f16x4_to_f32x4_fp16(v: &[u16; 4]) -> [f32; 4] {
     let mut vec = MaybeUninit::<uint16x4_t>::uninit();
     ptr::copy_nonoverlapping(v.as_ptr(), vec.as_mut_ptr().cast(), 4);
@@ -69,8 +75,9 @@ pub(super) unsafe fn f16x4_to_f32x4_fp16(v: &[u16; 4]) -> [f32; 4] {
     *(&result as *const float32x4_t).cast()
 }
 
-#[target_feature(enable = "fp16")]
 #[inline]
+#[rustversion::since(1.59)]
+#[target_feature(enable = "fp16")]
 pub(super) unsafe fn f32x4_to_f16x4_fp16(v: &[f32; 4]) -> [u16; 4] {
     let mut vec = MaybeUninit::<float32x4_t>::uninit();
     ptr::copy_nonoverlapping(v.as_ptr(), vec.as_mut_ptr().cast(), 4);
@@ -83,8 +90,9 @@ pub(super) unsafe fn f32x4_to_f16x4_fp16(v: &[f32; 4]) -> [u16; 4] {
     *(&result as *const uint16x4_t).cast()
 }
 
-#[target_feature(enable = "fp16")]
 #[inline]
+#[rustversion::since(1.59)]
+#[target_feature(enable = "fp16")]
 pub(super) unsafe fn f16x4_to_f64x4_fp16(v: &[u16; 4]) -> [f64; 4] {
     let mut vec = MaybeUninit::<uint16x4_t>::uninit();
     ptr::copy_nonoverlapping(v.as_ptr(), vec.as_mut_ptr().cast(), 4);
@@ -102,8 +110,9 @@ pub(super) unsafe fn f16x4_to_f64x4_fp16(v: &[u16; 4]) -> [f64; 4] {
     *[low, high].as_ptr().cast()
 }
 
-#[target_feature(enable = "fp16")]
 #[inline]
+#[rustversion::since(1.59)]
+#[target_feature(enable = "fp16")]
 pub(super) unsafe fn f64x4_to_f16x4_fp16(v: &[f64; 4]) -> [u16; 4] {
     let mut low = MaybeUninit::<float64x2_t>::uninit();
     let mut high = MaybeUninit::<float64x2_t>::uninit();
@@ -122,8 +131,9 @@ pub(super) unsafe fn f64x4_to_f16x4_fp16(v: &[f64; 4]) -> [u16; 4] {
     *(&result as *const uint16x4_t).cast()
 }
 
-#[target_feature(enable = "fp16")]
 #[inline]
+#[rustversion::since(1.59)]
+#[target_feature(enable = "fp16")]
 pub(super) unsafe fn add_f16_fp16(a: u16, b: u16) -> u16 {
     let result: u16;
     asm!(
@@ -135,8 +145,9 @@ pub(super) unsafe fn add_f16_fp16(a: u16, b: u16) -> u16 {
     result
 }
 
-#[target_feature(enable = "fp16")]
 #[inline]
+#[rustversion::since(1.59)]
+#[target_feature(enable = "fp16")]
 pub(super) unsafe fn subtract_f16_fp16(a: u16, b: u16) -> u16 {
     let result: u16;
     asm!(
@@ -148,8 +159,9 @@ pub(super) unsafe fn subtract_f16_fp16(a: u16, b: u16) -> u16 {
     result
 }
 
-#[target_feature(enable = "fp16")]
 #[inline]
+#[rustversion::since(1.59)]
+#[target_feature(enable = "fp16")]
 pub(super) unsafe fn multiply_f16_fp16(a: u16, b: u16) -> u16 {
     let result: u16;
     asm!(
@@ -161,8 +173,9 @@ pub(super) unsafe fn multiply_f16_fp16(a: u16, b: u16) -> u16 {
     result
 }
 
-#[target_feature(enable = "fp16")]
 #[inline]
+#[rustversion::since(1.59)]
+#[target_feature(enable = "fp16")]
 pub(super) unsafe fn divide_f16_fp16(a: u16, b: u16) -> u16 {
     let result: u16;
     asm!(
